Updates for completed restrictors.
Completed restrictors now also won't show unless the calculator is in simple wish planning mode and the end banner is the current banner or if its in advanced wishing mode and the current banner is selected for at least one of the banner end dropdowns.

Fixed issues with their triggers.

Updated the trigger input and change for all inputs to now include selects and moved it into an on document ready.