Import/Export upgrades.
Combined import and export into a single accordion.

Made export into a button that will copy straight to clip board. This button will also run the update export state functionality so the function was removed.

Replaced the scroll parameter in the import function with a run parameter and so now the import function will know whether to scroll or call the calculate function based on that param.