Numerical code improvements.
Renamed Lost5050s to LostCharacter5050s

Added ModifiedWishPlanMapper so we don't have to keep checking if wish plan items are active.

WishGroupMaxWishes will now be initially set to the first wish groups limit so we don't need the First check. We will also use this variable to avoid having to index the limit.

The wish sims will now return true or false, for if they were successful or not, instead of the number of wish items won.

The wish sim code will now use the variables FiveStarChance and NonFiveStarChance to check if a five star has been won. NonFiveStarChance will be multiplied by the chance of not getting a five star, every time we make a wish, and once it is less than FiveStarChance we will pass the check.

Cleaned up the capturing radiance win rate code.